+++ src/main/java/ca/keal/sastrane/chess/ChessAI.java
@@ -14,16 +14,15 @@ import java.util.Map;
 
 public class ChessAI extends AI {
     
-    // TODO migrate AI to double instead of int
-    private static final Map<Class<? extends Piece>, Integer> pieceToNaiveValue = new HashMap<>();
+    private static final Map<Class<? extends Piece>, Double> pieceToNaiveValue = new HashMap<>();
     
     static {
-        pieceToNaiveValue.put(Pawn.class, 1);
-        pieceToNaiveValue.put(Knight.class, 3);
-        pieceToNaiveValue.put(Bishop.class, 3);
-        pieceToNaiveValue.put(Rook.class, 5);
-        pieceToNaiveValue.put(Queen.class, 9);
-        pieceToNaiveValue.put(King.class, (int) 1e9); // not Integer.MAX_VALUE to avoid overflow
+        pieceToNaiveValue.put(Pawn.class, 1.0);
+        pieceToNaiveValue.put(Knight.class, 3.0);
+        pieceToNaiveValue.put(Bishop.class, 3.0);
+        pieceToNaiveValue.put(Rook.class, 5.0);
+        pieceToNaiveValue.put(Queen.class, 9.0);
+        pieceToNaiveValue.put(King.class, 1e9); // not Double.MAX_VALUE to avoid overflow
     }
     
     public ChessAI(double difficulty) {
@@ -31,23 +30,23 @@ public class ChessAI extends AI {
     }
     
     @Override
-    public int heuristic(@NonNull Round round, @NonNull Player player) {
+    public double heuristic(@NonNull Round round, @NonNull Player player) {
         // draw => MIN_VALUE/2, win => MAX_VALUE, lose => MIN_VALUE, else: naive piece value method
         // TODO improve piece valuation over naive method
         Result result = Chess.getInstance().getResult(round);
         if (result instanceof Result.Win) {
-            return ((Result.Win) result).getPlayer() == player ? Integer.MAX_VALUE : Integer.MIN_VALUE;
+            return ((Result.Win) result).getPlayer() == player ? Double.MAX_VALUE : Double.MIN_VALUE;
         } else if (result == Result.DRAW) {
-            return Integer.MIN_VALUE / 2;
+            return Double.MIN_VALUE / 2;
         }
         
         // value = sum of own pieces - sum of everyone else's pieces
-        int sumOwnPieces = 0;
-        int sumOpponentPieces = 0;
+        double sumOwnPieces = 0;
+        double sumOpponentPieces = 0;
         for (Square square : round.getBoard()) {
             Pair<Piece, Player> atSquare = round.getBoard().get(square);
             if (atSquare != null) {
-                int naiveValue = pieceToNaiveValue.get(atSquare.getLeft().getClass());
+                double naiveValue = pieceToNaiveValue.get(atSquare.getLeft().getClass());
                 if (atSquare.getRight() == player) {
                     sumOwnPieces += naiveValue;
                 } else {
